# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HRMS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      parameters:
        - name: departmentId
          in: query
          schema: { type: ["integer", "null"] }
        - name: designationId
          in: query
          schema: { type: ["integer", "null"] }
        - name: status
          in: query
          schema: { type: ["string", "null"] }
        - name: search
          in: query
          schema: { type: ["string", "null"] }
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    patch:
      operationId: updateEmployee
      tags: [employees]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── DEPARTMENTS ───────────────────────────────────────
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      parameters:
        - name: employeeId
          in: query
          schema: { type: ["integer", "null"] }
        - name: status
          in: query
          schema: { type: ["string", "null"] }
        - name: month
          in: query
          schema: { type: ["string", "null"] }
      responses:
        "200":
          description: Expense claims
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}/approve:
    patch:
      operationId: approveExpense
      tags: [expenses]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"

  # ── VENDORS ───────────────────────────────────────────
...
components:
  schemas:
...
    Employee:
      type: object
      required: [id, employeeId, firstName, lastName, email, phone, role, status, joiningDate]
      properties:
        id: { type: integer }
        employeeId: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        phone: { type: string }
        role: { type: string }
        status: { type: string }
        joiningDate: { type: string }
        departmentId: { type: ["integer", "null"] }
        departmentName: { type: ["string", "null"] }
        designationId: { type: ["integer", "null"] }
        designationName: { type: ["string", "null"] }
        managerId: { type: ["integer", "null"] }
        managerName: { type: ["string", "null"] }
        bankAccount: { type: ["string", "null"] }
        ifscCode: { type: ["string", "null"] }
        bankName: { type: ["string", "null"] }
        panNumber: { type: ["string", "null"] }
        aadharNumber: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        emergencyContact: { type: ["string", "null"] }
        profilePhoto: { type: ["string", "null"] }
        createdAt: { type: string }
    EmployeeInput:
      type: object
      required: [firstName, lastName, email, phone, role, joiningDate]
      properties:
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        phone: { type: string }
        role: { type: string }
        joiningDate: { type: string }
        departmentId: { type: ["integer", "null"] }
        designationId: { type: ["integer", "null"] }
        managerId: { type: ["integer", "null"] }
        bankAccount: { type: ["string", "null"] }
        ifscCode: { type: ["string", "null"] }
        bankName: { type: ["string", "null"] }
        panNumber: { type: ["string", "null"] }
        aadharNumber: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        emergencyContact: { type: ["string", "null"] }
        profilePhoto: { type: ["string", "null"] }
    EmployeeUpdate:
      type: object
      properties:
        firstName: { type: ["string", "null"] }
        lastName: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        role: { type: ["string", "null"] }
        status: { type: ["string", "null"] }
        departmentId: { type: ["integer", "null"] }
        designationId: { type: ["integer", "null"] }
        managerId: { type: ["integer", "null"] }
        bankAccount: { type: ["string", "null"] }
        ifscCode: { type: ["string", "null"] }
        bankName: { type: ["string", "null"] }
        panNumber: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        emergencyContact: { type: ["string", "null"] }
        profilePhoto: { type: ["string", "null"] }

    # ── DEPARTMENT ────────────────────────────────────
...
    Expense:
      type: object
      required: [id, employeeId, category, amount, date, status]
      properties:
        id: { type: integer }
        employeeId: { type: integer }
        employeeName: { type: ["string", "null"] }
        category: { type: string }
        amount: { type: number }
        date: { type: string }
        description: { type: ["string", "null"] }
        billPhoto: { type: ["string", "null"] }
        status: { type: string }
        managerApproval: { type: ["string", "null"] }
        accountsApproval: { type: ["string", "null"] }
        isAutoTravel: { type: boolean }
        travelKm: { type: ["number", "null"] }
        travelRate: { type: ["number", "null"] }
        approvedBy: { type: ["integer", "null"] }
        rejectionReason: { type: ["string", "null"] }
        createdAt: { type: string }
    ExpenseInput:
      type: object
      required: [employeeId, category, amount, date]
      properties:
        employeeId: { type: integer }
        category: { type: string }
        amount: { type: number }
        date: { type: string }
        description: { type: ["string", "null"] }
        billPhoto: { type: ["string", "null"] }
        isAutoTravel: { type: ["boolean", "null"] }
        travelKm: { type: ["number", "null"] }
        travelRate: { type: ["number", "null"] }

    # ── VENDOR ────────────────────────────────────────
...
    ApprovalInput:
      type: object
      required: [action]
      properties:
        action: { type: string }
        remarks: { type: ["string", "null"] }
        approvedBy: { type: ["integer", "null"] }

    # ── OFFICE LOCATION ───────────────────────────────
...