# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HRMS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Employee {
  id: number;
  employeeId: string;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  role: string;
  status: string;
  joiningDate: string;
  /** @nullable */
  departmentId?: number | null;
  /** @nullable */
  departmentName?: string | null;
  /** @nullable */
  designationId?: number | null;
  /** @nullable */
  designationName?: string | null;
  /** @nullable */
  managerId?: number | null;
  /** @nullable */
  managerName?: string | null;
  /** @nullable */
  bankAccount?: string | null;
  /** @nullable */
  ifscCode?: string | null;
  /** @nullable */
  bankName?: string | null;
  /** @nullable */
  panNumber?: string | null;
  /** @nullable */
  aadharNumber?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  emergencyContact?: string | null;
  /** @nullable */
  profilePhoto?: string | null;
  createdAt?: string;
}