# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HRMS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: number;
  employeeId: number;
  /** @nullable */
  employeeName?: string | null;
  category: string;
  amount: number;
  date: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  billPhoto?: string | null;
  status: string;
  /** @nullable */
  managerApproval?: string | null;
  /** @nullable */
  accountsApproval?: string | null;
  isAutoTravel?: boolean;
  /** @nullable */
  travelKm?: number | null;
  /** @nullable */
  travelRate?: number | null;
  /** @nullable */
  approvedBy?: number | null;
  /** @nullable */
  rejectionReason?: string | null;
  createdAt?: string;
}